+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## X.Y.Z (INSERT_DATE_HERE)
 
+#### Minor Changes
+- [Unresolved subrule references can cause none descriptive exceptions during follow up validations.](https://github.com/SAP/chevrotain/issues/209)
+- [More descriptive unresolved grammar error message.](https://github.com/SAP/chevrotain/issues/210)
+
 #### Documentation
 - [Dynamically defined Tokens example.](https://github.com/SAP/chevrotain/issues/208)
 
@@ -8,7 +12,7 @@
 ## 0.11.1 (6-16-2016)
 
 #### Minor Changes
-   - [extendToken should create constructor functions with meaningful names](https://github.com/SAP/chevrotain/issues/206)
+- [extendToken should create constructor functions with meaningful names](https://github.com/SAP/chevrotain/issues/206)
 
 
 
@@ -17,7 +21,7 @@
 - Minor version release as the previous version included a breaking change in a patch version.
  
 #### Minor Changes
-   - [Better runtime checks for Invalid Lexer definitions.](https://github.com/SAP/chevrotain/issues/204)
+- [Better runtime checks for Invalid Lexer definitions.](https://github.com/SAP/chevrotain/issues/204)